+++ docs/changelog.md
@@ -71,6 +71,10 @@
 
 ## 更新说明详情
 
+1.12.76 fix: fix bond_zh_hs_spot interface
+
+    1. 修复 bond_zh_hs_spot 接口
+
 1.12.75 add: add futures_contract_info_ine interface
 
     1. 新增 futures_contract_info_ine 接口
@@ -3623,6 +3627,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.12.76 fix: fix bond_zh_hs_spot interface
 
 1.12.75 add: add futures_contract_info_ine interface